+++ CHANGELOG.md
@@ -21,6 +21,7 @@ attribute, which is used for container domain name in NNS contracts (#2954)
 - Unathorized container ops accepted by the IR (#2947)
 - Structure table in the SN-configuration document (#2974)
 - False negative connection to NeoFS chain in multi-endpoint setup with at least one live node (#2986)
+- Overriding the default container and object attributes only with the appropriate flags (#2985)
 
 ### Changed
 - `ObjectService`'s `Put` RPC handler caches up to 10K lists of per-object sorted container nodes (#2901)